added tag/product tag columns

+++ develop/models/Product.js
@@ -35,7 +35,7 @@ Product.init(
       }
     },
     category_id: {
-      type: DataTypes.INTEGER(10),
+      type: DataTypes.INTEGER,
       allowNull: false,
       references: {
         model: 'category',
